Refactor HeatMapColumn and HeatMapContainer widgets
The HeatMapColumn widget has been refactored to include a new dataValue variable that holds the value of the dataset for a specific date. This change allows us to pass this value directly to the HeatMapContainer.

In addition, the key parameter in both widgets' constructors has been changed from an optional named parameter to a positional one using super.key.

The HeatMapContainer widget now includes an OptionalTooltip that displays the dataValue when hovered over. The dataValue is also added as a new parameter in its constructor.

Lastly, some minor changes have been made to improve code readability and structure.